Rename fields index file to base, so we can create an entry point via a new index file for loading "base", "container" and package fields.
